Revert "specify '3.0 (quilt)' source format"
This reverts commit 34fa27e8d6282aee1e2134b07bdae0f6930c5ee8.

Turns out that debian-x is in two worlds at the same time:

We use the orig tarball to build the package,
but we directly track upstream in git in the packaging.

source format 1.0 is better than 3.0 (quilt) for that workflow. With
3.0 (quilt), as an example, the following fails when upstream has
differences between the git repository and the released tarball (and
upstream does, in our case):

git clone git://git.debian.org/git/pkg-xorg/driver/xserver-xorg-video-intel
cd xserver-xorg-video-intel
uscan --download-current-version
debuild -S